Make new module for error types
